+++ Changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added Presentation, Documents and ItemDecision to ApiAgendaItem
+
 ### Changed
 - Refactored the method FirstAgendaService::getCommitteesInOrganizations and 
   FirstAgendaService::getAllCommitteesAvailable to use private method.